split-utils * Move time-functions to utils/time.ts
